requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
decorator==4.4.2
future==0.18.2
idna==2.9
iniconfig==1.0.1
loguru==0.5.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.0.1
python-dotenv==0.13.0
python-telegram-bot==12.7
requests==2.23.0
six==1.15.0
toml==0.10.1
tornado==6.0.4
urllib3==1.25.9
win32-setctime==1.0.1
